feat: implement location form for adding and editing locations with validation and error handling
